+++ radar-upload-backend/src/test/java/org/radarbase/upload/resource/RecordResourceTest.kt
@@ -27,7 +27,6 @@ import org.radarbase.upload.Config
 import org.radarbase.upload.api.RecordDTO
 import org.radarbase.upload.api.RecordDataDTO
 import org.radarbase.upload.api.SourceTypeDTO
-import org.radarbase.upload.logger
 import org.radarbase.upload.mock.MockResourceEnhancerFactory
 import org.radarcns.auth.token.RadarToken
 import java.io.IOException
